Conflicts:
	source/javaclient/com/ib/client/EReader.java

+++ samples/Cpp/TestCppClientSSL/TestCppClientSSL.cpp
@@ -171,8 +171,8 @@ void TestCppClient::cancelOrder()
 
 ///////////////////////////////////////////////////////////////////
 // events
-void TestCppClient::orderStatus( OrderId orderId, const std::string& status, int filled,
-	                            int remaining, double avgFillPrice, int permId, int parentId,
+void TestCppClient::orderStatus( OrderId orderId, const std::string& status, double filled,
+	                            double remaining, double avgFillPrice, int permId, int parentId,
 	                            double lastFillPrice, int clientId, const std::string& whyHeld)
 
 {
@@ -231,7 +231,7 @@ void TestCppClient::winError( const std::string& str, int lastError) {}
 void TestCppClient::connectionClosed() {}
 void TestCppClient::updateAccountValue(const std::string& key, const std::string& val,
                                        const std::string& currency, const std::string& accountName) {}
-void TestCppClient::updatePortfolio(const Contract& contract, int position,
+void TestCppClient::updatePortfolio(const Contract& contract, double position,
                                     double marketPrice, double marketValue, double averageCost,
                                     double unrealizedPNL, double realizedPNL, const std::string& accountName){}
 void TestCppClient::updateAccountTime(const std::string& timeStamp) {}
@@ -263,7 +263,7 @@ void TestCppClient::deltaNeutralValidation(int reqId, const UnderComp& underComp
 void TestCppClient::tickSnapshotEnd(int reqId) {}
 void TestCppClient::marketDataType(TickerId reqId, int marketDataType) {}
 void TestCppClient::commissionReport( const CommissionReport& commissionReport) {}
-void TestCppClient::position( const std::string& account, const Contract& contract, int position, double avgCost) {}
+void TestCppClient::position( const std::string& account, const Contract& contract, double position, double avgCost) {}
 void TestCppClient::positionEnd() {}
 void TestCppClient::accountSummary( int reqId, const std::string& account, const std::string& tag, const std::string& value, const std::string& curency) {}
 void TestCppClient::accountSummaryEnd( int reqId) {}
