+++ samples/Java/samples/rfq/SimpleWrapper.java
@@ -13,7 +13,6 @@ import com.ib.client.CommissionReport;
 import com.ib.client.Contract;
 import com.ib.client.ContractDetails;
 import com.ib.client.DeltaNeutralContract;
-import com.ib.client.EClient;
 import com.ib.client.EClientSocket;
 import com.ib.client.EJavaSignal;
 import com.ib.client.EReader;
@@ -141,7 +140,7 @@ public class SimpleWrapper implements EWrapper {
 		logIn("tickEFP");
 	}
 
-	public void orderStatus(int orderId, String status, int filled, int remaining,
+	public void orderStatus(int orderId, String status, double filled, double remaining,
 			double avgFillPrice, int permId, int parentId, double lastFillPrice,
 			int clientId, String whyHeld) {
 		logIn("orderStatus");    	
@@ -159,7 +158,7 @@ public class SimpleWrapper implements EWrapper {
 		logIn("updateAccountValue");
 	}
 
-	public void updatePortfolio(Contract contract, int position, double marketPrice, double marketValue,
+	public void updatePortfolio(Contract contract, double position, double marketPrice, double marketValue,
 			double averageCost, double unrealizedPNL, double realizedPNL, String accountName) {
 		logIn("updatePortfolio");
 	}
@@ -261,7 +260,7 @@ public class SimpleWrapper implements EWrapper {
 	}
 
 	
-	public void position(String account, Contract contract, int pos, double avgCost) {
+	public void position(String account, Contract contract, double pos, double avgCost) {
 		logIn("position");
 	}
